Move footer to bottom of page and make it sticky
Initially, the footer followed directly on from the search area.
However, given the main content of the page is short, the footer did not
sit at the bottom of the browser window, meaning there was a large
amount of space after the footer unless the browser was very narrow,
which isn't visually appealing.

This commit updates the footer so that it sits at the bottom of the
browser window, and as the browser is resized the footer will move with
it until it reaches the main content, at which point the footer will
stop and not overlap the rest of the page content, i.e. it is 'sticky'.

This approach ensures the footer always appears at the bottom of the
browser, rather than hard coding its position, and ensures the page
still works as a whole at smaller browser heights by not hiding the main
content.