Refactor classes to improve readability
Previously, elements such as <footer> and certain structual <divs> had
unnecessary or poorly named classes which made it more difficult to
comprehend the page structure.

This commit removes the redundant classes, and adds their styling to HTML
elements, and updates the names of other classes to be more
descriptive/appropriate, thereby improving the readability of the page
structure/styling.