Fix outline focus style
Initially, the default browser :focus outline styles for the search
field in Chrome & Firefox were removed by setting no outline style on
all elements when in focus.

This isn't a good solution, as it unnecessarily targets all elements
rather than the specific one being affected.

This commit updates this behaviour by removing the universal selector
and instead applying the relevant rule to the specific areas of the site
that were affected.