Refactor ID's and classes
Previously, index.html contained elements with a mixture of IDs and
classes , both of which had styling applied to them.

This commit refactors the code so that only classes are used to style
elements, rather than IDs and classes, in order to improve the
maintainability of the code and avoid issues around ID specificity.